build_assets/gcp_run_python_dockerfiles/requirements.grpc.txt
tensorflow==2.11.1
tensorflow-serving-api==2.11.1
protobuf==3.19.6
grpcio==1.54.2
requests